Very rough start
Defined a spring boot application that does nothing. Eventually it will
e either a command line interface for generating the files or a maven
plugin.

Defined a grammar and got it building in the pom but can't find the
generated classes. They may just go to target at the moment.